+++ lib/src/controls/navigation/navigation_view/view.dart
@@ -56,18 +56,14 @@ class NavigationView extends StatefulWidget {
 
   /// Can be used to override the widget that is built from
   /// the [PaneItem.body]. Only used if [pane] is provided.
-  /// If nothing is selected, `selectedIndex` will be -1 and
-  /// `selectedPaneItemBody` will be null. If you specify this
-  /// function, you would typically return a [NavigationBody],
-  /// passing `ValueKey(selectedIndex ?? -1)` as the
-  /// [NavigationBody.itemKey] and `viewWidget.transitionBuilder` as
-  /// the [NavigationBody.transitionBuilder].
+  /// If nothing is selected, `selectedPaneItemBody` will be
+  /// null.
   ///
   /// This can be useful if you are using router-based navigation,
-  /// and the body of the navigation pane should be determined by
-  /// the current route rather than just by the currently selected pane.
-  final Widget Function(NavigationView viewWidget, int? selectedIndex,
-      Widget? selectedPaneItemBody)? paneBodyBuilder;
+  /// and the body of the navigation pane is dynamically determined or
+  /// affected by the current route rather than just by the currently
+  /// selected pane.
+  final Widget Function(Widget? selectedPaneItemBody)? paneBodyBuilder;
 
   /// The navigation pane, that can be displayed either on the
   /// left, on the top, or above the body.
@@ -345,16 +341,16 @@ class NavigationViewState extends State<NavigationView> {
       if (widget.pane != null) {
         final pane = widget.pane!;
         final paneBodyBuilder = widget.paneBodyBuilder;
-        final body = paneBodyBuilder != null
-            ? paneBodyBuilder(widget, pane.selected ?? -1,
-                pane.selected != null ? pane.selectedItem.body : null)
-            : NavigationBody(
-                itemKey: ValueKey(pane.selected ?? -1),
-                transitionBuilder: widget.transitionBuilder,
-                child: pane.selected != null
-                    ? pane.selectedItem.body
-                    : const SizedBox.shrink(),
-              );
+        final body = _NavigationBody(
+          itemKey: ValueKey(pane.selected ?? -1),
+          transitionBuilder: widget.transitionBuilder,
+          child: paneBodyBuilder != null
+              ? paneBodyBuilder(
+                  pane.selected != null ? pane.selectedItem.body : null)
+              : (pane.selected != null
+                  ? pane.selectedItem.body
+                  : const SizedBox.shrink()),
+        );
         if (pane.customPane != null) {
           paneResult = Builder(builder: (context) {
             return PaneScrollConfiguration(